package: move partner package to community package
