+++ libs/community/langchain_community/embeddings/naver.py
