+++ libs/community/tests/integration_tests/chat_models/test_naver.py
@@ -1,5 +1,5 @@
 """Test ChatNaver chat model."""
-from langchain_naver.chat_models import ChatClovaX
+from langchain_community.chat_models.naver import ChatClovaX
 
 
 def test_stream() -> None:
