+++ libs/community/tests/integration_tests/embeddings/test_naver.py
@@ -1,5 +1,5 @@
 """Test Naver embeddings."""
-from langchain_naver.embeddings import ClovaXEmbeddings
+from langchain_community.embeddings.naver import ClovaXEmbeddings
 
 
 def test_langchain_naver_embedding_documents() -> None:
