--- libs/partners/naver/.gitignore
@@ -1,1 +1,0 @@
-__pycache__
