--- libs/partners/naver/scripts/check_imports.py
@@ -1,17 +1,0 @@
-import sys
-import traceback
-from importlib.machinery import SourceFileLoader
-
-if __name__ == "__main__":
-    files = sys.argv[1:]
-    has_failure = False
-    for file in files:
-        try:
-            SourceFileLoader("x", file).load_module()
-        except Exception:
-            has_faillure = True
-            print(file)  # noqa: T201
-            traceback.print_exc()
-            print()  # noqa: T201
-
-    sys.exit(1 if has_failure else 0)
