--- libs/partners/naver/tests/unit_tests/test_imports.py
@@ -1,10 +1,0 @@
-from langchain_naver import __all__
-
-EXPECTED_ALL = [
-    "ChatNaver",
-    "ClovaXEmbeddings"
-]
-
-
-def test_all_imports() -> None:
-    assert sorted(EXPECTED_ALL) == sorted(__all__)
